refactor(server): consolidate `home_flag` logic into `main.go` for better maintainability and simplicity
